add library.js and variable.js file to the modules
