+++ index.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -7,58 +8,50 @@
     <title>Awesome Books</title>
     <link rel="stylesheet" href="index.css">
 </head>
+
 <body>
     <!-- navbar -->
     <nav id="navbar" class="navbar">
-
-            <div class="web-name">
-                <p>Awesome Books</p>
-            </div>
-            
-            <div class="web-menu">
-    
-                <ul class="menu-list">
-                    <li class="menu-list-items"><a href="#" id="books-list-view" class="active">List</a></li>
-                    <li class="menu-list-items"><a href="#" id="add-book-view">Add New</a></li>
-                    <li class="menu-list-items"><a href="#" id="contact-us-view">Contact</a></li>
-                </ul>
-    
+        <p class="web-name">Awesome Books</p>
+        <div class="web-menu">
+            <a href="#" id="books-list-view" class="active">List</a>
+            <a href="#" id="add-book-view">Add New</a>
+            <a href="#" id="contact-us-view">Contact</a>
         </div>
-
     </nav>
-    <!-- navbar closed -->
+
+    <!-- book list section -->
     <section id="books-list">
         <div class="container">
             <h1 id="all">All awesome books</h1>
-        <div class="book-list">
-            <!-- See JS File -->
-        </div>
+            <div class="book-list">
+                <!-- dynamically in js -->
+            </div>
         </div>
     </section>
 
-        <section id="add-book">
-            <div class="container">
-                <h1 id="new">Add a new book</h1>
-        <form action="" method="" id="form">
-            <div>
-                <input type="text" name="titleInput" id="input_title" class="form-control" placeholder="Title"/>
-            </div>
-            <br/>
-            <div> 
-                <input type="text" name="authorInput" id="input_author" class="form-control" placeholder="Author"/>
-            </div>
-            <br/>
-            <button id="add_btn">Add</button>
-        </form>
-            </div>
-        </section>
+    <!-- add book section -->
+    <section id="add-book">
+        <div class="container">
+            <h1 id="new">Add a new book</h1>
+            <form action="" method="" id="form">
+                <input type="text" name="titleInput" id="input_title" class="form-control" placeholder="Title"
+                    required />
+                <br />
+                <input type="text" name="authorInput" id="input_author" class="form-control" placeholder="Author"
+                    required />
+                <br />
+                <button id="add_btn" type="submit">Add Book</button>
+            </form>
+        </div>
+    </section>
+
     <!-- contact us section -->
     <section id="contact-us">
         <div class="container">
             <h1>Contact Information</h1>
-        <div class="contact-info">
-            <p>
-                Lorem ipsum dolor, sit amet consectetur adipisicing elit. Dolorum cumque consequatur aut iusto reprehenderit provident assumenda, sed dolores quaerat temporibus velit quas voluptatibus iste veniam nobis saepe nam pariatur nisi!
+            <p class="contact-info">
+                Do you have any questions or you just want to say "Hello"? <br>You can reach out to us!
             </p>
             <ul class="contact-details">
                 <li>Our email is abc123@gmail.com</li>
@@ -66,10 +59,9 @@
                 <li>Our address is Helphome, Help city, Help</li>
             </ul>
         </div>
-        </div>
     </section>
-    <!-- contact us section closed -->
 
+    <!-- footer section-->
     <footer id="copyright">
         <span>Copyright</span>
         <span>...</span>
@@ -77,4 +69,5 @@
 
     <script src="index.js" type="module"></script>
 </body>
+
 </html>
