+++ index.js
@@ -1,47 +1,9 @@
-import Books from './modules/book.js';
+import Library from './modules/library.js';
+import { form, bookList, titleInput, authorInput, addBtn, booksListView, addNewBookView, contactView, listbook, addbook, contactus } from './modules/variable.js';
 
-// menu buttons closed
-const form = document.querySelector('#form');
-const bookList = document.querySelector('.book-list');
-const titleInput = document.querySelector('#input_title');
-const authorInput = document.querySelector('#input_author');
-const addBtn = document.querySelector('#add_btn');
-
-class Library {
-  constructor() {
-    this.Library = JSON.parse(localStorage.getItem('collections')) || [];
-  }
-
-  addBook() {
-    const newBook = new Books();
-    newBook.title = titleInput.value;
-    newBook.author = authorInput.value;
-    this.Library.push(newBook);
-    localStorage.setItem('collections', JSON.stringify(this.Library));
-    this.showBook();
-  }
-
-  showBook() {
-    bookList.innerHTML = '';
-    this.Library.forEach((collection, i) => {
-      bookList.innerHTML += `<div id="${i}" class="books">
-            <p class="book-info">
-                <span class="title">"${collection.title}" by ${collection.author}</span>
-            </p>
-            <button class="remove-btn">Remove</button>
-    </div>`;
-    });
-  }
-
-  removeBook(book) {
-    book.parentElement.remove();
-    this.Library = this.Library.filter((collection, i) => i !== Number(book.parentElement.id));
-    localStorage.setItem('collections', JSON.stringify(this.Library));
-  }
-}
-
+// book list
 window.onload = () => {
-  const newLibrary = new Library();
+  let newLibrary = new Library();
   newLibrary.showBook();
 
   form.addEventListener('submit', (e) => {
@@ -60,24 +22,14 @@ window.onload = () => {
 
   bookList.addEventListener('click', (e) => {
     if (e.target.className.includes('remove-btn')) {
-      const data = e.target;
+      let data = e.target;
       newLibrary.removeBook(data);
     }
     window.location.reload();
   });
 };
 
-// calling all views
-// menu buttons
-const booksListView = document.querySelector('#books-list-view');
-const addNewBookView = document.querySelector('#add-book-view');
-const contactView = document.querySelector('#contact-us-view');
-
-// calling section
-const listbook = document.querySelector('#books-list');
-const addbook = document.querySelector('#add-book');
-const contactus = document.querySelector('#contact-us');
-
+// changing contents
 booksListView.onclick = () => {
   booksListView.classList.add('active');
   contactView.classList.remove('active');
