+++ modules/variable.js
@@ -1,0 +1,15 @@
+const form = document.querySelector('#form');
+const bookList = document.querySelector('.book-list');
+const titleInput = document.querySelector('#input_title');
+const authorInput = document.querySelector('#input_author');
+const addBtn = document.querySelector('#add_btn');
+
+const booksListView = document.querySelector('#books-list-view');
+const addNewBookView = document.querySelector('#add-book-view');
+const contactView = document.querySelector('#contact-us-view');
+
+const listbook = document.querySelector('#books-list');
+const addbook = document.querySelector('#add-book');
+const contactus = document.querySelector('#contact-us');
+
+export { form, bookList, titleInput, authorInput, addBtn, booksListView, addNewBookView, contactView, listbook, addbook, contactus };
